feat(leadspace): move leadspace to patterns package
