--- packages/patterns-react/src/patterns/LeadSpace/README.md
@@ -1,37 +1,0 @@
-# Lead Space
-
-> The Lead Space pattern is to be utilized within IBM.com for Lorem Ipsum.
-
-## Getting started
-
-Here's a quick example to get you started.
-
-```javascript
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { LeadSpace } from '@carbon/ibmdotcom-patterns-react';
-import '@carbon/ibmdotcom-styles/scss/patterns/leadspace/index.scss';
-function App() {
-  return <LeadSpace />;
-}
-ReactDOM.render(<App />, document.querySelector('#app'));
-```
-
-> 💡 Don't forget to import the horizontal rule styles from
-> [@carbon/ibmdotcom-styles](/packages/styles).
-
-## Stable selectors
-
-| Name        | Description |
-| ----------- | ----------- |
-| `leadspace` | Pattern     |
-
-## 🙌 Contributing
-
-We're always looking for contributors to help us fix bugs, build new features,
-or help us improve the project documentation. If you're interested, definitely
-check out our [Contributing Guide](/.github/CONTRIBUTING.md)! 👀
-
-## 📝 License
-
-Licensed under the [Apache 2.0 License](/LICENSE).
